Update login page, add new images, and modify CSS styles
