--- App.css
@@ -1,220 +1,0 @@
-/* src/ForgotPassword.css */
-body{
-  margin: 0;
-  padding: 0;
-  background-color: #ffffff;
- 
-    background-image: url('last1.jpg');
-    background-repeat: no-repeat;
-    background-position: left center;
-    background-size: cover;
-  }
-  
-
-/* Styling for full-screen layout */
-.full-screen {
-  height: 100vh; /* Full height of the viewport */
-  width:100vw;
-  display:block; /* Use flexbox for layout */
-  align-items: center; /* Center items vertically */
-  justify-content: center; /* Center items horizontally */
-  align-self: center; /* Center align the form */
-  background-size:cover;
-}
-
-
-/* Styling for the forgot password form */
-.form {
-  width: 500px; 
-  max-width: 100vw; 
-  padding: 20px; 
-  display: flex; 
-  flex-direction: column; 
-  align-items: center; 
-  align-self: center; 
-  height: 60%;
-  max-height: 100vh;
-  justify-content: center; 
-  border-radius: 10px;
-  box-shadow: 0px 6px 6px -3px rgba(0,0,0,0.2), 0px 10px 14px 1px rgba(0,0,0,0.14), 0px 4px 18px 3px rgba(0,0,0,0.12);
-  
-}
-
-/* Styling for forgot password input fields */
-.forgot-password-input {
-  width: 100vw; /* Full width */
-  margin-bottom: 15px; /* Bottom margin for spacing */
-}
-
-/* Styling for forgot password buttons */
-
-
-/* Styling for button container */
-.button-container {
-  display: flex; /* Use flexbox for layout */
-  align-self: center;
-  justify-content: center; /* Add space between buttons */
-  margin-top: 15px; /* Top margin for spacing */
-  margin-left: 0%;
-  width: 50vw; /* Full width */
-  
-  
-}
-
-/* Styling for text color */
-.typoColor {
-  color: #7b1fa2; /* Purple color for text */
-  text-align: center; /* Center-align text */
-  font-size: 10em; /* Adjust as needed */
-  font-weight: bold;
-}
-
-/* src/Login.css */
-
-/* Styling for login container */
-.container {
-  display: flex; 
-  justify-content: center; 
-  align-items: center; 
-  height: 100%; 
-  width: 100vw; 
-  max-width: 100vw;
-  /*background-color: #ae1184; */
- 
-}
-
-.left {
-  background-image: url("finalback.jpg");
-  background-size: cover;
-  width: 50%;
-}
-
-.right {
-  width: 50%;
-}
-
-/* Styling for card content in login */
-/*.card-content {
-  width: 100vw; /* Full width of the viewport
-  max-width: 100vw; /* Maximum width, missing 'px' - check for correctness 
-  padding: 20px; /* Padding around the card content 
-  box-sizing: border-box; /* Include padding in width calculation 
-}*/
-
-/* Styling for second login container */
-.container2 {
-  display: flex; /* Use flexbox for layout */
-  flex-direction: column; /* Stack child elements vertically */
-  align-items: center; /* Center items horizontally */
-  align-self: center;
-  justify-content: center; /* Center items vertically */
-  width: 100vw; /* Full width of the viewport */
-  height: 100vh; /* Full height of the viewport */
-  background-color: #f5f5f5; /* Background color */
-  padding: 20px; /* Padding around the container */
-}
-
-/* Styling for links in login */
-.link-instance {
-  align-items: flex !important;  
-  align-self: flex-end !important;
-  display: flex !important; 
-  flex: 0 0 auto !important; 
-  justify-content: flex-end !important; 
-  width: 100vw !important; 
-  text-decoration: none; 
-}
-
-.link-instance:hover {
-  text-decoration: underline; /* Underline on hover */
-}
-
-.link3 {
-  align-items: center !important; /* Center items vertically, important to override styles if needed */
-  display: flex !important; /* Use flexbox for layout, important to override styles if needed */
-  flex: 0 0 auto !important; /* Non-flexible, auto-sizing, important to override styles if needed */
-  justify-content: center !important; /* Center items horizontally, important to override styles if needed */
-  width: 520px !important; /* Fixed width, important to override styles if needed */
-  text-decoration: none; /* Remove default text decoration for links */
-}
-
-.link3:hover {
-  text-decoration: underline; /* Underline on hover */
-}
-
-/* Styling for buttons in login */
-.button-instance {
-  align-self: center !important; /* Stretch to full width, important to override styles if needed */
-  box-shadow: 0px 3px 1px -2px #00000033 !important; /* Box shadow effect, important to override styles if needed */
-  display: flex !important; /* Use flexbox for layout, important to override styles if needed */
-  flex: 0 0 auto !important; /* Non-flexible, auto-sizing, important to override styles if needed */
-  width: 25% !important; /* Full width of the parent container, important to override styles if needed */
-}
-
-/* Styling for text wrapper */
-.span {
-  align-items: center; /* Center items vertically */
-  align-self: center; /* Stretch to full width */
-  display: flex; /* Use flexbox for layout */
-  flex-direction: column; /* Stack child elements vertically */
-  gap: 4px; /* Gap between child elements */
-  height: 50px; /* Fixed height */
-  justify-content: center; /* Center items vertically */
-  position: relative; /* Relative positioning */
-  width: 100vw; /* Full width of the parent container */
-}
-
-/* Styling for text color and font in login */
-.text-wrapper-2 {
-  color: #7b1fa2; /* Purple color for text */
-  font-family: var(--typography-body1-font-family); /* Font family */
-  font-size: var(--typography-body1-font-size); /* Font size */
-  font-style: var(--typography-body1-font-style); /* Font style */
-  font-weight: var(--typography-body1-font-weight); /* Font weight */
-  letter-spacing: var(--typography-body1-letter-spacing); /* Letter spacing */
-  line-height: var(--typography-body1-line-height); /* Line height */
-  margin-top: -2px; /* Top margin for spacing, negative value */
-  position: relative; /* Relative positioning */
-  white-space: nowrap; /* No wrapping for text */
-  width: fit-content; /* Adjust width based on content */
-}
-
-/* Styling for label color */
-.label-color {
-  color: #7b1fa2; /* Purple color for text */
-}
-
-/* Styling for center alignment */
-.center {
-  display: flex; /* Use flexbox for layout */
-  align-items: center; /* Center items vertically */
-  justify-content: center; /* Center items horizontally */
-  height: 100vh; /* Full height of the viewport */
-  width:100vw;
-}
-
-.textfiledStyle {
-  align-self: center !important; /* Stretch to full width, important to override styles if needed */
-  justify-self: center !important; /* Stretch to full width, important to override styles if needed */
-  display: flex !important; /* Use flexbox for layout, important to override styles if needed */
-  flex: 0 0 auto !important; /* Non-flexible, auto-sizing, important to override styles if needed */
-  width: 75% !important; /* Full width of the parent container, important to override styles if needed */
-}
-
-.textfiledStyle1 {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.button-instance:first-child {
-  margin-right: 10px;
-}
-
-.logo {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 50%; /* Adjust as needed */
-}
-
