--- AppPages.tsx
@@ -1,6 +1,0 @@
-// src/AppPages.tsx
-import LoginPage from './LoginPage';
-import PasswordResetPage from './PasswordResetPage';
-//import './App.css';
-
-export { LoginPage, PasswordResetPage };
