+++ src/App.css
@@ -1,5 +1,5 @@
 #root {
-	max-width: 1280px;
+	max-width: 100vw;
 	margin: 0 auto;
 	text-align: center;
 	color: black; /* TODO: replace & resolve with mui theme when it's finalized */
