+++ src/styles/style.css
@@ -1,0 +1,140 @@
+.FM-login {
+    background-color: #ffffff;
+    height: 1024px;
+    position: relative;
+    width: 1440px;
+  }
+  
+  .FM-login .text-field-instance {
+    border-radius: 30px !important;
+    height: 53px !important;
+    left: 824px !important;
+    position: absolute !important;
+    top: 398px !important;
+    width: 360px !important;
+  }
+  
+  .FM-login .design-component-instance-node {
+    align-self: unset !important;
+    width: 210px !important;
+  }
+  
+  .FM-login .text-field-2 {
+    margin-bottom: -3px !important;
+  }
+  
+  .FM-login .text-field-3 {
+    text-shadow: 0px 4px 4px #00000040 !important;
+  }
+  
+  .FM-login .text-field-4 {
+    border-radius: 30px !important;
+    height: 53px !important;
+    left: 824px !important;
+    position: absolute !important;
+    top: 485px !important;
+    width: 360px !important;
+  }
+  
+  .FM-login .div-wrapper {
+    align-items: center;
+    display: flex;
+    gap: 8px;
+    height: 40px;
+    justify-content: center;
+    left: 943px;
+    padding: 10px 12px;
+    position: absolute;
+    top: 655px;
+    width: 119px;
+  }
+  
+  .FM-login .label-text-3 {
+    color: var(--m-3syslightprimary);
+    font-family: var(--m3-label-large-font-family);
+    font-size: var(--m3-label-large-font-size);
+    font-style: var(--m3-label-large-font-style);
+    font-weight: var(--m3-label-large-font-weight);
+    letter-spacing: var(--m3-label-large-letter-spacing);
+    line-height: var(--m3-label-large-line-height);
+    margin-left: -6.5px;
+    margin-right: -6.5px;
+    margin-top: -1px;
+    position: relative;
+    text-align: center;
+    white-space: nowrap;
+    width: fit-content;
+  }
+  
+  .FM-login .p {
+    color: #000000;
+    font-family: var(--m3-title-medium-font-family);
+    font-size: var(--m3-title-medium-font-size);
+    font-style: var(--m3-title-medium-font-style);
+    font-weight: var(--m3-title-medium-font-weight);
+    height: 24px;
+    left: 929px;
+    letter-spacing: var(--m3-title-medium-letter-spacing);
+    line-height: var(--m3-title-medium-line-height);
+    position: absolute;
+    text-align: center;
+    top: 781px;
+    width: 150px;
+  }
+  
+  .FM-login .text-wrapper-2 {
+    color: #000000;
+    font-family: var(--m3-display-medium-font-family);
+    font-size: var(--m3-display-medium-font-size);
+    font-style: var(--m3-display-medium-font-style);
+    font-weight: var(--m3-display-medium-font-weight);
+    height: 52px;
+    left: 939px;
+    letter-spacing: var(--m3-display-medium-letter-spacing);
+    line-height: var(--m3-display-medium-line-height);
+    position: absolute;
+    text-align: center;
+    top: 278px;
+    width: 114px;
+  }
+  
+  .FM-login .button-instance {
+    display: flex !important;
+    left: 954px !important;
+    position: absolute !important;
+    top: 605px !important;
+    width: 92px !important;
+  }
+  
+  .FM-login .button-2 {
+    margin-left: -2.5px !important;
+    margin-right: -2.5px !important;
+  }
+  
+  .FM-login .button-3 {
+    display: flex !important;
+    left: 956px !important;
+    position: absolute !important;
+    top: 840px !important;
+    width: 92px !important;
+  }
+  
+  .FM-login .image {
+    height: 1024px;
+    left: 0;
+    object-fit: cover;
+    position: absolute;
+    top: 0;
+    width: 536px;
+  }
+  
+  .FM-login .logo-white {
+    background-blend-mode: difference;
+    height: 51px;
+    left: 839px;
+    object-fit: cover;
+    position: absolute;
+    top: 103px;
+    width: 336px;
+  }
+
